[triton 3.3] cpp_wrapper: add a global_scratch arg
Following triton # 4916, the generated cubin expects a global_scratch argument to support on-device TMA. We believe this is the source of many of the "invalid argument" failures on AOTI/cpp_wrapper tests. AFAIK, we don't use on-device TMA in Inductor as of now, so it should be safe to use a nullptr for the scratch space.

ghstack-source-id: 7abf0bd57665df8f459b5c402d1688424a00c883
Pull Request resolved: https://github.com/pytorch/pytorch/pull/148051