Don't crash when we call __qualname__ on torch._C.ScriptFunction
We've root caused this to correctly throwing attribute error on ScriptFunction
when missing attributes are caused. This PR will fix crashes that are showing
up. I'm going to stack a second PR to fix torch._c.ScriptFunction just being a
very badly behaving python object (which should also fix this

ghstack-source-id: 679f1a6b59ba3e98e5001fa811abb9444306ceef
Pull Request resolved: https://github.com/pytorch/pytorch/pull/147894