scriptfunction: Make sure we have valid __name__ and __qualname__
It's not fully clear why these are not being created, but you can definitely
reproduce this in code. __name__ is fun, since there appears to be no way to
explicitly set it on the pybind11 layer or c++ layer. I've set this in the
python wrapper code (which works correctly). But let me know if people feel
strongly and want us to go explicitly cast to python within the cpp functions
and set it there.

ghstack-source-id: 363aa9d05e400ff8044ce80bc7d4e2abf870c5ae
Pull Request resolved: https://github.com/pytorch/pytorch/pull/147906